refactor: correct components structure
