+++ src/pages/Anvil/components/AnvilButton/AnvilButton.tsx
@@ -1,13 +1,13 @@
 import { Button, ButtonProps } from '@mui/material';
 
-import { anvilCalcStyles } from '../styles';
+import { styles } from './styles';
 
 interface AnvulButtonProps extends ButtonProps {
     text: string;
 }
 
 export function AnvilButton(props: AnvulButtonProps) {
-    const classes = anvilCalcStyles();
+    const classes = styles();
 
     return (
         <Button {...props} variant='contained' color='secondary' className={classes.anvilCalcButton}>
