+++ src/pages/Anvil/components/AnvilButton/styles.ts
@@ -1,0 +1,9 @@
+import { createStyles, makeStyles } from '@mui/styles';
+
+export const styles = makeStyles(() =>
+    createStyles({
+        anvilCalcButton: {
+            fontSize: 48
+        }
+    })
+);
