+++ src/pages/Anvil/components/AnvilSection/styles.ts
@@ -1,0 +1,16 @@
+import { createStyles, makeStyles } from '@mui/styles';
+
+export const styles = makeStyles(() =>
+    createStyles({
+        anvilItem: {
+            width: 260
+        },
+        anvilItemInfo: {
+            color: '#fff',
+            '& span': {
+                fontWeight: 600,
+                color: '#fbf941'
+            }
+        }
+    })
+);
